+++ Modules/Ruleset.UserInfo/Ruleset.UserInfo.psm1
@@ -59,6 +59,7 @@ $PublicScripts = @(
 	"Get-UserGroup"
 	"Merge-SDDL"
 	"Split-Principal"
+	"Test-Credential"
 	"Test-UPN"
 )