+++ Templates/PipelineScript.ps1
@@ -94,7 +94,7 @@ begin
 {
 	#region Initialization
 	# TODO: Adjust path to project settings
-	. $PSScriptRoot\..\Config\ProjectSettings.ps1
+	. $PSScriptRoot\..\Config\ProjectSettings.ps1 $PSCmdlet
 	New-Variable -Name ThisScript -Scope Private -Option Constant -Value ((Get-Item $PSCommandPath).Basename)
 
 	# Check requirements