+++ Templates/TestRule.ps1
@@ -61,7 +61,7 @@ param (
 
 #region Initialization
 # TODO: adjust path to project settings
-. $PSScriptRoot\..\..\Config\ProjectSettings.ps1
+. $PSScriptRoot\..\..\Config\ProjectSettings.ps1 $PSCmdlet
 New-Variable -Name ThisScript -Scope Private -Option Constant -Value ((Get-Item $PSCommandPath).Basename)
 
 # Check requirements